refactor: move mapper to designated package
