+++ local/src/main/java/com/hxl/local/mapper/coin/CoinEntityMapper.java
@@ -1,4 +1,4 @@
-package com.hxl.local.mapper;
+package com.hxl.local.mapper.coin;
 
 import androidx.annotation.NonNull;
